+++ lib/exq_ui_web/live/queue_live/show.ex
@@ -14,7 +14,7 @@ defmodule ExqUIWeb.QueueLive.Show do
       ])
       |> assign(:actions, [
         %{name: "delete", label: "Delete"},
-        %{name: "delete_all", label: "Delete All"}
+        %{name: "delete_all", label: "Delete All", confirm: "Are you sure you want to delete all jobs in this queue? This action cannot be undone."}
       ])
       |> assign(:config, config)
